+++ example/lib/logic/driving_logic.dart
@@ -1,12 +1,23 @@
 import 'package:popsicle/popsicle.dart';
 
-class DrivingLogic extends PopsicleState<String> {
-  DrivingLogic() : super(state: '');
-  void check(int age) {
-    if (age >= 18) {
-      shift('Yes you can Drive');
-    } else {
-      shift('You are not eligable');
-    }
-  }
-}
+// class DrivingLogic extends Popsicle<String> {
+//   DrivingLogic() : super(state: '');
+
+//   void check(int age) {
+//     if (age >= 18) {
+//       shift('Yes you can Drive');
+//     } else {
+//       shift('You are not eligible');
+//     }
+//   }
+
+//   @override
+//   void onInit() {
+//     print('DrivingLogic initialized!');
+//   }
+
+//   @override
+//   void onDispose() {
+//     print('DrivingLogic disposed!');
+//   }
+// }