+++ lib/popsicle.dart
@@ -1,6 +1,7 @@
 library;
 
 import 'dart:async';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 part 'src/state/popsicle_state.dart';
@@ -12,3 +13,6 @@ part 'src/extensions/pop_extensions.dart';
 part 'src/ui/popsicle_ui.dart';
 part 'src/locator/popsicle_locator.dart';
 part 'src/locator/locator_provider.dart';
+part 'src/extensions/middlewares.dart';
+part 'src/extensions/logic_extension.dart';
+part 'src/state/logic_mixin.dart';