+++ lib/src/extensions/middlewares.dart
@@ -1,0 +1,29 @@
+part of 'package:popsicle/popsicle.dart';
+
+extension PopsicleStateMiddlewareExtension<T> on _BasePopsicleState<T> {
+  /// Add middleware easily
+  void addMiddleware(PopsicleMiddlewareBase<T> middleware) {
+    use((oldState, newState) => middleware(oldState, newState));
+  }
+}
+
+/// Popsicle middleware base class
+abstract class PopsicleMiddlewareBase<T> {
+  /// Transform or block updates
+  T? call(T oldState, T newState);
+}
+
+/// Example: logging middleware
+// class LoggingMiddleware<T> extends PopsicleMiddlewareBase<T> {
+//   @override
+//   T? call(T oldState, T newState) {
+//     if (kDebugMode) {
+//       print('Middleware: $oldState -> $newState');
+//     }
+//     return newState; // pass along
+//   }
+// }
+
+extension Popsicles on Type {
+  static T of<T extends Object>() => Popsicle.get<T>();
+}